+++ src/main/java/ca/corefacility/bioinformatics/irida/repositories/joins/project/ProjectSampleJoinRepository.java
@@ -2,6 +2,8 @@ package ca.corefacility.bioinformatics.irida.repositories.joins.project;
 
 import java.util.List;
 
+import javax.persistence.Tuple;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -95,4 +97,13 @@ public interface ProjectSampleJoinRepository
 	@Query("SELECT j.sample.id FROM ProjectSampleJoin j where j.owner=false AND j.project=?1")
 	public List<Long> getLockedSamplesForProject(Project project);
 
+	/**
+	 * Calculate the coverage for a list of Samples within a Project.
+	 * 
+	 * @param project   {@link Project} to get sample coverage for.
+	 * @param sampleIds {@link Sample} ids
+	 * @return A List of {@link Tuple}s with first entry as the sample id and second entry as the coverage
+	 */
+	@Query("SELECT ps.sample.id, ROUND(SUM(qc.totalBases)/ps.project.genomeSize) FROM ProjectSampleJoin ps join SampleSequencingObjectJoin sso on sso.sample = ps.sample join CoverageQCEntry qc on qc.sequencingObject = sso.sequencingObject where ps.project=?1 and ps.sample.id in ?2 group by ps.sample.id")
+	public List<Tuple> calculateCoverageForSamplesInProject(Project project, List<Long> sampleIds);
 }