+++ src/main/webapp/resources/js/pages/projects/samples/components/SamplesTable.jsx
@@ -244,7 +244,7 @@ export function SamplesTable() {
       title: i18n("SamplesTable.Column.sampleName"),
       className: "t-td-name",
       dataIndex: ["sample", "sampleName"],
-      sorter: { multiple: 1 },
+      sorter: true,
       render: (name, row) => (
         <a href={`${sampleUrl}/${row.sample.id}`}>{name}</a>
       ),
@@ -257,21 +257,29 @@ export function SamplesTable() {
     {
       title: i18n("SamplesTable.Column.quality"),
       width: 100,
-      dataIndex: "quality",
-      render: (qualities) => <SampleQuality qualities={qualities} />,
+      dataIndex: "qcStatus",
+      render: (qcStatus, row) => (
+        <SampleQuality qcStatus={qcStatus} qualities={row.quality} />
+      ),
+    },
+    {
+      title: i18n("SamplesTable.Column.coverage"),
+      className: "t-td-coverage",
+      width: 100,
+      dataIndex: "coverage",
     },
     {
       title: i18n("SamplesTable.Column.organism"),
       className: "t-td-organism",
       dataIndex: ["sample", "organism"],
-      sorter: { multiple: 1 },
+      sorter: true,
       ...getColumnSearchProps(["sample", "organism"], "t-organism-select"),
     },
     {
       title: i18n("SamplesTable.Column.project"),
       className: "t-td-project",
       dataIndex: ["project", "name"],
-      sorter: { multiple: 1 },
+      sorter: true,
       key: "associated",
       render: (name, row) => {
         if (!(row.project.id in projectColours)) {
@@ -300,14 +308,14 @@ export function SamplesTable() {
     {
       title: i18n("SamplesTable.Column.collectedBy"),
       dataIndex: ["sample", "collectedBy"],
-      sorter: { multiple: 1 },
+      sorter: true,
       ...getColumnSearchProps(["sample", "collectedBy"]),
     },
     {
       title: i18n("SamplesTable.Column.created"),
       className: "t-td-created",
       dataIndex: ["sample", "createdDate"],
-      sorter: { multiple: 1 },
+      sorter: true,
       width: 230,
       render: (createdDate) => {
         return formatInternationalizedDateTime(createdDate);
@@ -319,7 +327,7 @@ export function SamplesTable() {
       className: "t-td-modified",
       dataIndex: ["sample", "modifiedDate"],
       defaultSortOrder: "descend",
-      sorter: { multiple: 1 },
+      sorter: true,
       width: 230,
       render: (modifiedDate) => {
         return formatInternationalizedDateTime(modifiedDate);